Fix regression in the login job unit tests caused by r1130337. While still retaining its main feature which was to unbreak connection in the SSL/TLS (actually explains why the logs were empty in such cases).
Indeed that's the LoginJob which triggers the SSL support on the socket
in such cases. So it means the LoginJob has to be started right away at
socket connection time in such case... but not in the other ones
(unencrypted or STARTTLS) for which we have to wait for the server
greeting first.

Means I had to add a small heuristic for that in the Session class,
adding an internal dep on LoginJob. Not overly happy with it but does
the job.

I guess we might still have a potential issue in the SSL/TLS + PREAUTH
case, but I'm yet to see such a thing in practice...

CCMAIL: adam@kde.org
CCMAIL: asouza@kde.org

svn path=/trunk/KDE/kdepimlibs/; revision=1131096