Don't wait indefinitely for a session thread to terminate.
This just covers a bug a few layers up, where for some reason the IMAP
resource attempts a reconnect for no good reason and fails to stop the
IDLE session which hangs in an sync read which does not terminate itself
since the socket is still working perfectly fine. Nevertheless, it doesn't
hurt to have a sane fallback in case something goes wrong.

svn path=/trunk/KDE/kdepimlibs/; revision=1168823