automatic fallback for SSL/TLS protocol versions
Unfortunately some servers are to dump to fall back to the highest
supported version when the connection is initiated with e.g. an
TLS 1.2 ClientHello. This happens with openSSL 1.0.1 on the client side.

The server setup dialog is affected as well and will return encryption
unsupported for the server, this is not fixed yet.

BUG: 306964
REVIEW: 107099
FIXED-IN: 4.9.4
FIXED-IN: 4.10.0

Conflicts:
	kimap/sessionthread.cpp