Made the socketTimer a bit simpler and more robust.
There is no need for the code to not be a little more forgiving, and we can easily get into a situation
with the timer already running on startNext. e.g. due to socketActivity of due to unhandled data (BYE).

This will hopefully fix the ASSERT people were running into.

BUG: 310887