Fix crash when Session is destroyed while waiting for ignoreSslError()
ignoreSslError() is async and the whole Session including SessionThread
can be deleted in the meantime. Checking against member variable of the
potentially deleted object does not make much sense since the memory can
already be overwritten.

Instead we hold the pointer in a local QPointer variable, which will get
updated when the SessionThread is destroyed while we are waiting for
ignoreSslError() to return.

BUG: 332830
FIXED-IN: 16.04