StatusJob: remove pointless (and actually wrong) Q_ASSERT
The else branch would get evaluated when OK STATUS response
is parsed and it wrongly assumed that to be an error situation,
while that's a completely valid case.

Secondly we should not assert on (unexpected) responses from the
network, as any kinds of error might occur.

I'm wondering how nobody else is hitting this assert in production
and we are only seeing this on the CI.