Introduce MoveJob which implements the MOVE extension (RFC6851)
The MoveJob API is the same as CopyJob API, since the RFC6851
is modelled closely from the COPY command, having the same
parameters and responses.

Unlike with the traditional COPY + STORE \Deleted + EXPUNGE
approach, the RFC6851 guarantees the entire operation to be
atomic and is faster and uses less bandwidth.

When UIDPLUS extension is supported, the resulting UIDs are
returned via COPYUID response (like with CopyJob).