FetchJob: deprecate all signals, add a new one
Deprecate all the overloaded signals with many complicated arguments,
which makes connecting to them using the new connect() syntax (and
the old as well) tedious. Instead introduce KIMAP::Message structure
and a new messagesAvailable() signal that has a map of the Message
structures as the only argument. This will make the API much more
comfortable to use.

Eventually the deprecated signals should be removed in order to
simplify and speed up the response parsing code.