Fix STARTTLS support detection
Don't assume that IMAP servers announce capabilities in initial greeting,
the RFC does not require that at all. This assumption caused LoginJob
to fail if user configured STARTTLS but the server did not announce any
capabilities in the greeting.

Now we explicitly ask for capabilities before sending the STARTTLS command
to make sure that STARTTLS is supported.

BUG: 395249
FIXED-IN: 5.8.3